Add/Delete fully functional

+++ public/js/controllers/companies-controller.js
@@ -31,7 +31,7 @@ angular.module('app')
       }
 
       CompanyFactory.addCompany(name)
-        .then( function (res) {
+        .then( function (data) {
           $scope.name = '';
         })
         .catch( function(error) {
@@ -41,14 +41,21 @@ angular.module('app')
     };
 
     $scope.deleteCompany = function(id) {
-      console.log('---------Company Factory--------', CompanyFactory);
+      if ( !confirm('Are you sure you want to delete this company?') ) {
+        return;
+      }
+      //console.log(id);
       CompanyFactory.deleteCompany(id)
         .catch( function(error) {
           console.error(error);
         });
-
+      $scope.getAll();
     };
 
+    $scope.appliedToCompany = function() {
+
+    };
+    
     $scope.getAll();
 
   }]);
